Move AnonymousType to PHP\Types\Models
