+++ PHP/Collections/Collection/AnonymousKeyType.php
@@ -3,6 +3,7 @@ declare( strict_types = 1 );
 
 namespace PHP\Collections\Collection;
 
+use PHP\Types\Models\AnonymousType;
 use PHP\Types\Models\Type;
 
 /**
